refactor: isolate logging utility
